requirements.txt
langchain==0.1.20
langchain-openai==0.0.13
langgraph==0.0.64
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
scikit-learn==1.3.2
spacy==3.7.2
gliner==0.0.9
transformers==4.35.2
torch==2.0.0
